boot.s: complete boot block
does what a booter does: loads a kernel file from the
root directory and launches it. deals with the messy
business of trampolining from real mode to long mode.

the only things missing for the moment are video mode
selection and spinning up APs, both of which are just
delayed until the kernel is ready for them.

introduced the mkboot.c utility which installs the
boot block without disturbing an existing filesystem.

minor corrections/additions in include/sys/fs.h

+++ include/sys/fs.h
@@ -89,8 +89,8 @@ struct filsys
 };
 
 #define FS_SUPER_MAGIC      0xABE01E50
-#define FS_SUPER_MAGIC2     0x87CD          /* homage to OS-9/6809 */
-#define FS_BOOT_MAGIC       0xAA55          /* BIOS boot signature */
+#define FS_SUPER_MAGIC2     (short) 0x87CD  /* homage to OS-9/6809 */
+#define FS_BOOT_MAGIC       (short) 0xAA55  /* BIOS boot signature */
 
 /* compute the starting blocks for various regions of disk */
